+++ Objects/Objects.Revit/RevitBeam.cs
@@ -1,13 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using Objects.BuiltElements;
 
 namespace Objects.Revit
 {
-  public class RevitVerticalOpening : Opening
+  public class RevitBeam : Beam
   {
-    public Element host { get; set; }
-
+    public string family { get; set; }
     public Dictionary<string, object> parameters { get; set; }
   }
 }
