+++ Objects/Objects.Revit/RevitBrace.cs
@@ -1,12 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using Objects.BuiltElements;
 
 namespace Objects.Revit
 {
-  public class RevitFootprintRoof : Roof
+  public class RevitBrace : Beam
   {
-    public Level cutOffLevel { get; set; }
+    public string family { get; set; }
     public Dictionary<string, object> parameters { get; set; }
   }
 }
