+++ Objects/Objects.Revit/RevitCurves.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Text;
+using Objects.BuiltElements;
 
 namespace Objects.Revit
 {
