+++ Objects/Objects.Revit/RevitFloor.cs
@@ -1,12 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using Objects.BuiltElements;
 
 namespace Objects.Revit
 {
-  public class RevitBrace : Beam
+  public class RevitFloor : Floor
   {
-    public string family { get; set; }
     public Dictionary<string, object> parameters { get; set; }
   }
 }
