+++ Objects/Objects.Revit/RevitFootprintRoof.cs
@@ -1,13 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using Objects.BuiltElements;
 
 namespace Objects.Revit
 {
-  public class RevitWallOpening : Opening
+  public class RevitFootprintRoof : Roof
   {
-    public RevitWall host { get; set; }
-
+    public Level cutOffLevel { get; set; }
     public Dictionary<string, object> parameters { get; set; }
   }
 }
