+++ Objects/Objects.Revit/RevitLevel.cs
@@ -1,12 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using Objects.BuiltElements;
 
 namespace Objects.Revit
 {
-  public class RevitBeam : Beam
+  public class RevitLevel : Level
   {
-    public string family { get; set; }
+    public bool createView { get; set; }
     public Dictionary<string, object> parameters { get; set; }
   }
 }
