+++ Objects/Objects.Revit/RevitShaft.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using Objects.BuiltElements;
 
 namespace Objects.Revit
 {
