+++ Objects/Objects.Revit/RevitVerticalOpening.cs
@@ -1,12 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using Objects.BuiltElements;
 
 namespace Objects.Revit
 {
-  public class RevitLevel : Level
+  public class RevitVerticalOpening : Opening
   {
-    public bool createView { get; set; }
+    public Element host { get; set; }
+
     public Dictionary<string, object> parameters { get; set; }
   }
 }
