+++ Objects/Objects.sln
@@ -19,6 +19,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConverterRevit", "Converter
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConverterRevitTests", "Converters\ConverterRevit\ConverterRevitTests\ConverterRevitTests.csproj", "{D9C88135-B4B3-4C6C-A9D9-01A59A00BC0C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Objects.Revit", "Objects.Revit\Objects.Revit.csproj", "{51162368-2D07-4AEB-AB2D-984B86A4506C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -49,6 +51,10 @@ Global
 		{D9C88135-B4B3-4C6C-A9D9-01A59A00BC0C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D9C88135-B4B3-4C6C-A9D9-01A59A00BC0C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D9C88135-B4B3-4C6C-A9D9-01A59A00BC0C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{51162368-2D07-4AEB-AB2D-984B86A4506C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{51162368-2D07-4AEB-AB2D-984B86A4506C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{51162368-2D07-4AEB-AB2D-984B86A4506C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{51162368-2D07-4AEB-AB2D-984B86A4506C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
